권한체크

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
+import shop.mtcoding.blog._core.errors.exception.Exception404;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -31,7 +33,23 @@ public class BoardController {
 
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest req) {
-        Board board = boardRepository.findByIdJoinUser(id);
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = null;
+
+        try {
+            board = boardRepository.findByIdJoinUser(id);
+        } catch (Exception e) {
+            throw new Exception404("게시글을 찾을 수 없습니다.");
+        }
+
+        boolean isOwner = false;
+        if (sessionUser != null) {
+            if (sessionUser.getId() == board.getUser().getId()){
+                isOwner = true;
+            }
+        }
+
+        req.setAttribute("isOwner", isOwner);
         req.setAttribute("board",board);
         return "board/detail";
     }
@@ -45,6 +63,13 @@ public class BoardController {
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable Integer id) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 삭제 할 권한이 없습니다.");
+        }
+
         boardRepository.deleteById(id);
         return "redirect:/";
     }
@@ -52,12 +77,23 @@ public class BoardController {
     @GetMapping("/board/{id}/update-form")
     public String updateForm(@PathVariable Integer id, HttpServletRequest req) {
         Board board = boardRepository.findById(id);
+
+        if (board == null) {
+            throw new Exception404("해당 게시글을 찾을 수 없습니다.");
+        }
+
         req.setAttribute("board",board);
         return "/board/update-form";
     }
 
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id, BoardRequest.UpdateDTO reqDTO) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 수정 할 권한이 없습니다.");
+        }
         boardRepository.updateById(id, reqDTO.getTitle(), reqDTO.getContent());
         return "redirect:/board/" + id;
     }
